renamt the battery client to be https to make it clear it's not using mqtt
